refactor unload.
