+++ src/query/storages/stage/src/append/parquet_file/mod.rs
@@ -12,9 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#![allow(clippy::uninlined_format_args)]
-
-mod append;
-mod stage_table;
-
-pub use stage_table::StageTable;
+mod block_batch;
+mod limit_file_size_processor;
+mod pipeline;
+mod writer_processor;
+pub(crate) use pipeline::append_data_to_parquet_files;
