+++ src/query/storages/stage/src/append/parquet_file/pipeline.rs
@@ -19,18 +19,15 @@ use databend_common_catalog::table_context::TableContext;
 use databend_common_exception::Result;
 use databend_common_formats::FileFormatOptionsExt;
 use databend_common_pipeline_core::Pipeline;
-use databend_common_pipeline_sources::input_formats::InputContext;
 use opendal::Operator;
 
-use crate::row_based_file::limit_file_size_processor::LimitFileSizeProcessor;
-use crate::row_based_file::serialize_processor::SerializeProcessor;
-use crate::row_based_file::sink_processor::RowBasedFileSink;
+use super::limit_file_size_processor::LimitFileSizeProcessor;
+use super::writer_processor::ParquetFileWriter;
 
-// SerializeProcessor * N: serialize each data block to many small byte buffers.
-// LimitFileSizeProcessor * 1:  group small byte buffers to batches (as a block meta) that are large enough as a file.
-// RowBasedFileSink * N: simply concat small byte buffers to a whole and write out.
+// LimitFileSizeProcessor * 1:  slice/group block to batches (as a block meta) that are suitable as a file.
+// ParquetFileSink * N: simply serialize blocks in each meta to a whole file and write out.
 #[allow(clippy::too_many_arguments)]
-pub(crate) fn append_data_to_row_based_files(
+pub(crate) fn append_data_to_parquet_files(
     pipeline: &mut Pipeline,
     ctx: Arc<dyn TableContext>,
     table_info: StageTableInfo,
@@ -40,27 +37,6 @@ pub(crate) fn append_data_to_row_based_files(
     uuid: String,
     group_id: &std::sync::atomic::AtomicUsize,
 ) -> Result<()> {
-    let mut options_ext = FileFormatOptionsExt::create_from_settings(&ctx.get_settings(), false)?;
-    let output_format = options_ext.get_output_format(
-        table_info.schema(),
-        table_info.stage_info.file_format_params.clone(),
-    )?;
-    let compression = table_info
-        .stage_info
-        .file_format_params
-        .clone()
-        .compression();
-    let prefix = output_format.serialize_prefix()?;
-
-    pipeline.add_transform(|input, output| {
-        let mut options_ext =
-            FileFormatOptionsExt::create_from_settings(&ctx.get_settings(), false)?;
-        let output_format = options_ext.get_output_format(
-            table_info.schema(),
-            table_info.stage_info.file_format_params.clone(),
-        )?;
-        SerializeProcessor::try_create(ctx.clone(), input, output, output_format)
-    })?;
     pipeline.try_resize(1)?;
     pipeline.add_transform(|input, output| {
         LimitFileSizeProcessor::try_create(input, output, max_file_size)
@@ -68,19 +44,21 @@ pub(crate) fn append_data_to_row_based_files(
     if max_file_size != usize::MAX {
         pipeline.try_resize(max_threads)?;
     }
-
-    let compression = InputContext::get_compression_alg_copy(compression, "")?;
-
     pipeline.add_sink(|input| {
         let gid = group_id.fetch_add(1, std::sync::atomic::Ordering::Relaxed);
-        RowBasedFileSink::try_create(
+        let mut options_ext =
+            FileFormatOptionsExt::create_from_settings(&ctx.get_settings(), false)?;
+        let output_format = options_ext.get_output_format(
+            table_info.schema(),
+            table_info.stage_info.file_format_params.clone(),
+        )?;
+        ParquetFileWriter::try_create(
             input,
             table_info.clone(),
+            output_format,
             op.clone(),
-            prefix.clone(),
             uuid.clone(),
             gid,
-            compression,
         )
     })?;
     Ok(())
