+++ src/query/storages/stage/src/append/row_based_file/writer_processor.rs
@@ -18,61 +18,68 @@ use std::sync::Arc;
 
 use async_trait::async_trait;
 use databend_common_catalog::plan::StageTableInfo;
+use databend_common_compress::CompressAlgorithm;
+use databend_common_compress::CompressCodec;
 use databend_common_exception::Result;
 use databend_common_expression::BlockMetaInfoDowncast;
 use databend_common_expression::DataBlock;
-use databend_common_formats::output_format::OutputFormat;
 use databend_common_pipeline_core::processors::Event;
 use databend_common_pipeline_core::processors::InputPort;
 use databend_common_pipeline_core::processors::Processor;
 use databend_common_pipeline_core::processors::ProcessorPtr;
 use opendal::Operator;
 
-use crate::parquet_file::block_batch::BlockBatch;
-use crate::stage_table::unload_path;
+use super::buffers::FileOutputBuffers;
+use crate::append::path::unload_path;
 
-pub struct ParquetFileSink {
+pub struct RowBasedFileWriter {
     input: Arc<InputPort>,
     table_info: StageTableInfo,
-    output_format: Box<dyn OutputFormat>,
 
+    // always blocks for a whole file if not empty
     input_data: Option<DataBlock>,
+    // always the data for a whole file if not empty
     output_data: Vec<u8>,
 
     data_accessor: Operator,
+    prefix: Vec<u8>,
 
     uuid: String,
     group_id: usize,
     batch_id: usize,
+
+    compression: Option<CompressAlgorithm>,
 }
 
-impl ParquetFileSink {
+impl RowBasedFileWriter {
     pub fn try_create(
         input: Arc<InputPort>,
         table_info: StageTableInfo,
-        output_format: Box<dyn OutputFormat>,
         data_accessor: Operator,
+        prefix: Vec<u8>,
         uuid: String,
         group_id: usize,
+        compression: Option<CompressAlgorithm>,
     ) -> Result<ProcessorPtr> {
-        Ok(ProcessorPtr::create(Box::new(ParquetFileSink {
-            input,
+        Ok(ProcessorPtr::create(Box::new(RowBasedFileWriter {
             table_info,
-            output_format,
+            input,
             input_data: None,
-            output_data: vec![],
             data_accessor,
+            prefix,
             uuid,
             group_id,
             batch_id: 0,
+            output_data: vec![],
+            compression,
         })))
     }
 }
 
 #[async_trait]
-impl Processor for ParquetFileSink {
+impl Processor for RowBasedFileWriter {
     fn name(&self) -> String {
-        "ParquetFileSink".to_string()
+        "StageSink".to_string()
     }
 
     fn as_any(&mut self) -> &mut dyn Any {
@@ -102,23 +109,28 @@ impl Processor for ParquetFileSink {
     fn process(&mut self) -> Result<()> {
         let block = self.input_data.take().unwrap();
         let block_meta = block.get_owned_meta().unwrap();
-        let blocks = BlockBatch::downcast_from(block_meta).unwrap();
-        for b in blocks.blocks {
-            self.output_format.serialize_block(&b)?;
+        let buffers = FileOutputBuffers::downcast_from(block_meta).unwrap();
+        let size = buffers.buffers.iter().map(|b| b.len()).sum::<usize>();
+        let mut output = Vec::with_capacity(self.prefix.len() + size);
+        output.extend_from_slice(self.prefix.as_slice());
+        for b in buffers.buffers {
+            output.extend_from_slice(b.as_slice());
         }
-        self.output_data = self.output_format.finalize()?;
+        if let Some(compression) = self.compression {
+            output = CompressCodec::from(compression).compress_all(&output)?;
+        }
+        self.output_data = output;
         Ok(())
     }
 
     #[async_backtrace::framed]
     async fn async_process(&mut self) -> Result<()> {
-        assert!(!self.output_data.is_empty());
         let path = unload_path(
             &self.table_info,
             &self.uuid,
             self.group_id,
             self.batch_id,
-            None,
+            self.compression,
         );
         let data = mem::take(&mut self.output_data);
         self.data_accessor.write(&path, data).await?;
