+++ src/query/storages/stage/src/stage_table.rs
@@ -15,11 +15,9 @@
 use std::any::Any;
 use std::io::Read;
 use std::ops::Deref;
-use std::sync::atomic::AtomicUsize;
 use std::sync::Arc;
 
 use dashmap::DashMap;
-use databend_common_base::base::uuid;
 use databend_common_catalog::plan::DataSourceInfo;
 use databend_common_catalog::plan::DataSourcePlan;
 use databend_common_catalog::plan::PartInfo;
@@ -32,7 +30,6 @@ use databend_common_catalog::plan::StageTableInfo;
 use databend_common_catalog::table::AppendMode;
 use databend_common_catalog::table::Table;
 use databend_common_catalog::table_context::TableContext;
-use databend_common_compress::CompressAlgorithm;
 use databend_common_exception::ErrorCode;
 use databend_common_exception::Result;
 use databend_common_expression::BlockThresholds;
@@ -52,11 +49,9 @@ use opendal::Operator;
 use opendal::Scheme;
 use parking_lot::Mutex;
 
-use crate::parquet_file::append_data_to_parquet_files;
-use crate::row_based_file::append_data_to_row_based_files;
 /// TODO: we need to track the data metrics in stage table.
 pub struct StageTable {
-    table_info: StageTableInfo,
+    pub(crate) table_info: StageTableInfo,
     // This is no used but a placeholder.
     // But the Table trait need it:
     // fn get_table_info(&self) -> &TableInfo).
@@ -253,50 +248,7 @@ impl Table for StageTable {
         pipeline: &mut Pipeline,
         _: AppendMode,
     ) -> Result<()> {
-        let settings = ctx.get_settings();
-
-        let single = self.table_info.stage_info.copy_options.single;
-        let max_file_size = if single {
-            usize::MAX
-        } else {
-            let max_file_size = self.table_info.stage_info.copy_options.max_file_size;
-            if max_file_size == 0 {
-                // 256M per file by default.
-                256 * 1024 * 1024
-            } else {
-                let mem_limit = (settings.get_max_memory_usage()? / 2) as usize;
-                max_file_size.min(mem_limit)
-            }
-        };
-        let max_threads = settings.get_max_threads()? as usize;
-
-        let op = StageTable::get_op(&self.table_info.stage_info)?;
-        let fmt = self.table_info.stage_info.file_format_params.clone();
-        let uuid = uuid::Uuid::new_v4().to_string();
-        let group_id = AtomicUsize::new(0);
-        match fmt {
-            FileFormatParams::Parquet(_) => append_data_to_parquet_files(
-                pipeline,
-                ctx.clone(),
-                self.table_info.clone(),
-                op,
-                max_file_size,
-                max_threads,
-                uuid,
-                &group_id,
-            )?,
-            _ => append_data_to_row_based_files(
-                pipeline,
-                ctx.clone(),
-                self.table_info.clone(),
-                op,
-                max_file_size,
-                max_threads,
-                uuid,
-                &group_id,
-            )?,
-        };
-        Ok(())
+        self.do_append_data(ctx, pipeline)
     }
 
     // Truncate the stage file.
@@ -315,37 +267,5 @@ impl Table for StageTable {
     fn set_block_thresholds(&self, thresholds: BlockThresholds) {
         let mut guard = self.block_compact_threshold.lock();
         (*guard) = Some(thresholds)
-    }
-}
-
-pub fn unload_path(
-    stage_table_info: &StageTableInfo,
-    uuid: &str,
-    group_id: usize,
-    batch_id: usize,
-    compression: Option<CompressAlgorithm>,
-) -> String {
-    let format_name = format!(
-        "{:?}",
-        stage_table_info.stage_info.file_format_params.get_type()
-    )
-    .to_ascii_lowercase();
-
-    let suffix: &str = &compression
-        .map(|c| format!(".{}", c.extension()))
-        .unwrap_or_default();
-
-    let path = &stage_table_info.files_info.path;
-
-    if path.ends_with("data_") {
-        format!(
-            "{}{}_{:0>4}_{:0>8}.{}{}",
-            path, uuid, group_id, batch_id, format_name, suffix
-        )
-    } else {
-        format!(
-            "{}/data_{}_{:0>4}_{:0>8}.{}{}",
-            path, uuid, group_id, batch_id, format_name, suffix
-        )
     }
 }
